Neural Network from scratch/requirements.txt
numpy
flask
scikit-learn
pillow
